pyproject.toml
[tool.poetry]
name = "fulmo"
version = "0.1.0"
description = "Template to start your deep learning project based on `PyTorchLightning` for rapid prototyping."
authors = ["Gleb Glushkov <ptjexio@gmail.com>"]

readme = "README.md"
homepage = "https://github.com/jexio/fulmo"
repository = "https://github.com/jexio/fulmo"
keywords = ["hydra", "pytorch", "pytorch-lightning", "poetry", "template", "best-practices", "codestyle"]
classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    
    "License :: OSI Approved :: MIT License",
    
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/jexio/fulmo/issues"

[tool.poetry.dependencies]
python = "<3.10,>=3.8"
rich = ">=9.8.2,<11.0.0"
hydra-core = "1.1.0"
timm = "^0.4.5"
albumentations = ">=0.5.2,<1.1.0"
torch-optimizer = "^0.1.0"
imageio = "^2.9.0"
pandas = "^1.2.3"
nptyping = "^1.4.1"
scikit-image = "^0.18.1"
scikit-learn = "^0.24.2"
python-dotenv = ">=0.17.1,<0.20.0"
pytorch-lightning = "^1.3.6"
kornia = "^0.5.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
invoke = "^1.6.0"
flake8 = "^3.9.2"
flakehell = "^0.9.0"
flake8-builtins = "^1.5.3"
flake8-blind-except = "^0.2.0"
flake8-logging-format = "^0.6.0"
flake8-bugbear = "^21.4.3"
flake8-annotations = "^2.6.2"
flake8-docstrings = "^1.6.0"
flake8-bandit = "^2.1.2"
flake8-class-attributes-order = "^0.1.2"
interrogate = "^1.3.2"
darglint = "^1.8.0"
isort = "^5.9.3"
black = "^21.7b0"
safety = "^1.10.3"
mypy = "^0.910"
recommonmark = "^0.7.1"
pytest = "^6.2.4"
watchdog = {version = "^2.0.2", extras = ["watchmedo"]}
commitizen = "^2.17.13"

[tool.flakehell]
application-import-names = "fulmo"
format = "grouped"
max_line_length = 120
show_source = true
docstring-convention = "google"
class_attributes_order = ["field", "meta_class", "nested_class", "magic_method", "property_method", "static_method", "private_method", "method", "class_method"]
extended_default_ignore = []

[tool.flakehell.plugins]
pyflakes = ["+*"]
pycodestyle = ["+*", "-E203", "-E501"]
mccabe = ["+*"]
"flake8-*" = ["+*"]
flake8-bandit = ["+*", "-S101"]
flake8-docstrings = ["+*", "-D100", "-D104", "-D105", "-D107", "-D200", "-D204", "-D205", "-D301", "-D400", "-D401", "-D402", "-D412", "-D413"]
flake8-annotations = ["+*", "-ANN002", "-ANN003", "-ANN101", "-ANN102"]
darglint = ["+*"]

[tool.flakehell.exceptions."tests/"]
flake8-bandit = ["-S101"]

[tool.isort]
profile = "black"
atomic = true
include_trailing_comma = true
lines_after_imports = 2
use_parentheses = true
filter_files = true
line_length = 120
known_third_party = ["invoke", "nox"]

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
fail-under = 95
exclude = ["setup.py", "docs", "build", "*/*/__init__.py"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true

[tool.pytest]
norecursedirs = ["*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git __pycache__"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "fulmo/__version__.py",
    "pyproject.toml:version"
]

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
